src/main/java/diegojunco/springframework/didemo/controllers/PropertyInjectedController.java
package diegojunco.springframework.didemo.controllers;

import diegojunco.services.GreetingService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.beans.factory.annotation.Qualifier;
import org.springframework.stereotype.Controller;


/**
 * Created by Diego Junco on 15/03/2018
 **/
@Controller
public class PropertyInjectedController {

    @Autowired
    @Qualifier("greetingServiceImpl")
    public GreetingService greetingService;

    public String sayHello(){
        return greetingService.sayGreeting();
    }
}
